Move drop to own service
