+++ src/storage/lootAttribute.ts
@@ -5,13 +5,3 @@ import type { BotContext } from "@/context.js";
 import type { Database, Loot, LootId, LootInsertable, LootOrigin, LootTable } from "./db/model.js";
 
 import db from "@db";
-
-export interface LootAttribute {
-    id: number;
-    classId: number;
-    displayName: string;
-    shortDisplay: string | null;
-    color?: number;
-    initialDropWeight?: number;
-    visible: boolean;
-}
